Refactor roomba to move vacuum classes to separate folder
